requirements.txt
google-genai==1.28.0
python-dotenv==1.1.0
